git-svn-id: http://svn.purei.org/purei/trunk@164 b36208d7-6611-0410-8bec-b1987f11c4a2

+++ Source/CacheBlock.cpp
@@ -314,8 +314,8 @@ void CCacheBlock::StreamWrite(unsigned int nCount, ...)
 	
 	for(i = 0; i < nCount; i++)
 	{
-		nValue = va_arg(args, uint8);
-		StreamWriteByte(nValue);
+		nValue = va_arg(args, int);
+		StreamWriteByte(static_cast<uint8>(nValue));
 	}
 
 	va_end(args);
